requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.10
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
itsdangerous==2.2.0
blinker==1.9.0
colorama==0.4.6
